Move active-admin.js to utils.js
