+++ app/javascript/active_admin/lib/checkbox-toggler.js
@@ -1,4 +1,4 @@
-import ActiveAdmin from './active-admin';
+import ActiveAdmin from './utils';
 
 ActiveAdmin.CheckboxToggler = class CheckboxToggler {
   constructor(options, container){
