+++ app/javascript/active_admin/lib/per-page.js
@@ -1,4 +1,4 @@
-import ActiveAdmin from './active-admin';
+import ActiveAdmin from './utils';
 
 (($, ActiveAdmin) => {
