+++ app/javascript/active_admin/lib/utils.js
